+++ src/Screens/Home/HomeViewer.tsx
@@ -18,38 +18,16 @@ import { FloatingButton } from '~/Components/Atoms';
 import { getUniqueKey, stackNavigate } from '~/lib';
 
 interface HomeProps {
-  navigation: StackNavigationProp<LoginStackParams, 'Home'>;
+  sampleData: any;
+  handleNavigateToPostWrite: () => void;
+  handleNavigateSetParams: () => void;
 }
 
-type Type = 'normal' | 'best';
-const Home = ({ navigation }: HomeProps) => {
-  const sampleData = {
-    username: 'username',
-    createdAt: 'createdAt',
-    sentence: 'sentence',
-    heartCount: 45,
-    commentCount: 13,
-    type: 'best' as Type,
-  };
-  const handleNavigateToPostWrite = () => {
-    navigation.navigate('PostWrite');
-  };
-  const handleNavigateSetParams = () => {
-    navigation.navigate('PostDetail', {
-      heartCount: 40,
-      commentCount: 30,
-      username: '신입 가나다',
-      createdAt: '1시간전',
-      content: '소나무가 삐지면?\n칫솔',
-      isPressLike: true,
-      comments: [
-        {
-          username: '신입 카파하',
-          comment: '룰루룰루',
-        },
-      ],
-    });
-  };
+const HomeViewer = ({
+  sampleData,
+  handleNavigateToPostWrite,
+  handleNavigateSetParams,
+}: HomeProps) => {
   return (
     <BackgroundContainer>
       <StickyScrollView stickyPosition={1}>
@@ -61,11 +39,7 @@ const Home = ({ navigation }: HomeProps) => {
         </DeviceHeaderSticky>
         <DeviceSection>
           {Array.from({ length: 5 }, (_, index) => (
-            <PostCard
-              {...sampleData}
-              key={getUniqueKey(index)}
-              onPress={handleNavigateSetParams}
-            />
+            <PostCard {...sampleData} key={getUniqueKey(index)} onPress={handleNavigateSetParams} />
           ))}
         </DeviceSection>
       </StickyScrollView>
@@ -74,4 +48,4 @@ const Home = ({ navigation }: HomeProps) => {
   );
 };
 
-export default Home;
+export default HomeViewer;
