+++ src/Screens/Home/index.ts
@@ -1,0 +1,3 @@
+import HomeContainer from './HomeContainer';
+
+export default HomeContainer;
